requirements.txt
openai==1.1.1
python-dotenv==0.21.1 # For OpenAI auth
pyserial # For robot communication
openai-whisper
pydub
wavio
sounddevice